requirements.txt
fastapi
uvicorn
google-genai
dateparser
pydantic
python-multipart